requirements.txt
fastapi
uvicorn
torch
torchvision
facenet-pytorch
pillow
numpy